4.21*: bloglist expansion, step9

+++ controllers/blogs.js
@@ -45,18 +45,30 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   try {
-    
     const id = request.params.id;
-    const deletedBlog = await Blog.findByIdAndRemove(id)
+    const deletedBlog = await Blog.findById(id);
     
+    if (!deletedBlog) {
+      return response.status(404).json({ error: 'Blog not found' });
+    }
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' });
+    }
     
-    if (deletedBlog) {
-      response.status(204).end();
-    } 
+    if (deletedBlog.user.toString() !== decodedToken.id) {
+      return response.status(401).json({ error: 'unauthorized operation' });
+    }
+
+    await Blog.findByIdAndRemove(id)
+
+    response.status(204).end();
   } catch (error) {
-    response.status(404).json({ error: 'Blog not found' })
+    response.status(500).json({ error: 'An error occurred' });
   }
-})
+});
+
 
 blogsRouter.put('/:id', async (request, response) => {
   try {
